+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- Made `getRefsResolvingTo` able to get updated results when called inside an action (as long as the reference being back-traced is part of a root store).
 - Made `onInit`, `onAttachedToRootStore` and `onLazyInit` protected.
 
 ## 0.56.0